Extract game play route out of submissions.ts
Moves GET /api/games/:slug (serves a published game's assembled HTML, with
a fallback to the creator's own unpublished draft) and its cache/rate-limit
state into a new apps/api/src/catalog/game-play-route.ts module, alongside
catalog-routes.ts. Depends only on the already-extracted CatalogRoutesHandle
and DraftPreviewRoutesHandle, injected as options.

invalidatePublishedGameCaches stays in submissions.ts (also drives
catalogRoutes.invalidatePublishedGameCache) and now calls into the new
module's invalidateGameCache handle.

submissions.ts drops from 4323 to 4217 lines; the new file is 148 lines,
0 comment-prose debt.
