+++ eslint-rules/module-boundary-map.mjs
@@ -242,6 +242,7 @@ const FILE_BUCKET = {
 
   // catalog: github-client, snapshots, assemble, play
   'admin-game-routes': 'catalog',
+  'game-play-route': 'catalog',
   'music-tracks': 'catalog',
   'catalog-enricher': 'catalog',
   'catalog-vector-index': 'catalog',
